Create stub create task and get task endpoints
This is the first step in creating the tasks api. This ideally should be behind a feature flag as not to expose it to the end user.
It could be used to start a conversation with frontend developers. It could also be generated using some a library that generates data like faker.