Create a dependncy injection Container to be used with fastapi
This allows me to continue to flesh out the api, without having to worry about database setup, migrations, and integrating with a database.
I should be able to create a full api, all stored in memory before needing to set up the database.
I can focus on how the api will work, without having to consider what technologies I will use in the end. It will also allow me to easily change technologies and dependencies in the future.

By using the containers, we will be able to change what implementation of the TaskManager is used at runtime. I will use an InMemoryTaskManager for now.